update: 实现用户退出登录接口
reason: 无

+++ dao/user.go
@@ -7,14 +7,14 @@ import (
 
 type User struct{}
 
-func (*User) GetUserById(id int) (res entity.User) {
+func (*User) GetUserById(id uint64) (res entity.User) {
 	utils.DB.Table("user").
 		Where("id = ? AND status = 1", id).
 		First(&res)
 	return
 }
 
-func (*User) UpdateUserById(id int, updated entity.User) (res entity.User) {
+func (*User) UpdateUserById(id uint64, updated entity.User) (res entity.User) {
 	utils.DB.Table("user").
 		Where("id = ? AND status = 1", id).
 		Updates(updated).
